+++ R/ggtree.R
@@ -9,10 +9,10 @@
 ##' @param as.Date logical whether using Date class in time tree
 ##' @param yscale y scale
 ##' @param yscale_mapping yscale mapping for category variable
-##' @param ladderize logical (default \code{TRUE}). Should the tree be re-organized to have a 'ladder'
+##' @param ladderize logical (default `TRUE`). Should the tree be re-organized to have a 'ladder'
 ##' aspect?
-##' @param right logical. If \code{ladderize = TRUE}, should the ladder have the smallest clade on the
-##' right-hand side? See \code{\link[ape]{ladderize}} for more information. 
+##' @param right logical. If `ladderize = TRUE`, should the ladder have the smallest clade on the
+##' right-hand side? See [ape::ladderize()] for more information. 
 ##' @param branch.length variable for scaling branch, if 'none' draw cladogram
 ##' @param root.position position of the root node (default = 0)
 ##' @return tree
@@ -26,9 +26,17 @@
 ##' @importFrom ggplot2 coord_polar
 ##' @export
 ##' @author Yu Guangchuang
-##' @seealso \code{\link[ape]{ladderize}}
+##' @seealso [ape::ladderize()]
+##' @references 1. G Yu, TTY Lam, H Zhu, Y Guan (2018). Two methods for mapping and visualizing associated data
+##' on phylogeny using ggtree. Molecular Biology and Evolution, 35(2):3041-3043.
+##' <https://doi.org/10.1093/molbev/msy194>
+##'
+##' 2. G Yu, DK Smith, H Zhu, Y Guan, TTY Lam (2017). ggtree: an R package for
+##' visualization and annotation of phylogenetic trees with their covariates and
+##' other associated data. Methods in Ecology and Evolution, 8(1):28-36.
+##' <https://doi.org/10.1111/2041-210X.12628>
 ##' @examples
-##' require(ape)
+##' require(ape) 
 ##' tr <- rtree(10)
 ##' ggtree(tr)
 ggtree <- function(tr,
@@ -91,11 +99,11 @@ ggtree <- function(tr,
 
     if (layout == "circular" || layout == "radial") {
         p <- p + layout_circular()
-        ## refer to: https://github.com/GuangchuangYu/ggtree/issues/6
-        ## and also have some space for tree scale (legend)
-        p <- p + ylim(0, NA)
     } else if (layout == "fan") {
         p <- p + layout_fan(open.angle)
+    } else {
+        p <- p +
+            scale_y_continuous(expand = expand_scale(0, 0.6))
     }
 
     class(p) <- c("ggtree", class(p))